Applied Patch from Bug 10571: "Some Edit Boxes trigger an action on every user keystroke"
git-svn-id: svn://scribus.net/trunk/Scribus@17420 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/cpalette.h
@@ -84,8 +84,8 @@ public slots:
 	void displayColorValues(QString stroke, QString fille, int sShade, int fShade);
 	void displayGradient(int nr);
 	void displayOverprint(int);
-	void handleFillShade(int);
-	void handleStrokeShade(int);
+	void handleFillShade(double);
+	void handleStrokeShade(double);
 	void handleOverprint(int);
 	void handleFillGradient();
 	void handleStrokeGradient();
@@ -131,8 +131,8 @@ public slots:
 signals:
 	void NewPen(QString);
 	void NewBrush(QString);
-	void NewPenShade(int);
-	void NewBrushShade(int);
+	void NewPenShade(double);
+	void NewBrushShade(double);
 	void NewGradient(int);
 	void NewGradientS(int);
 	void NewPattern(QString);
